Remove project_manager.go entirely; tests passing
This is a major milestone. The old way of handling source information,
the projectManager, is totally gone, replaced by the new source-based
system. There's still work to be done there, but with tests passing,
we're more or less at parity with what we had before.
