+++ source.go
@@ -20,6 +20,20 @@ type sourceMetaCache struct {
 	// TODO(sdboyer) mutexes. actually probably just one, b/c complexity
 }
 
+// projectInfo holds manifest and lock
+type projectInfo struct {
+	Manifest
+	Lock
+}
+
+type existence struct {
+	// The existence levels for which a search/check has been performed
+	s sourceExistence
+
+	// The existence levels verified to be present through searching
+	f sourceExistence
+}
+
 func newMetaCache() *sourceMetaCache {
 	return &sourceMetaCache{
 		infos:  make(map[Revision]projectInfo),
